Changed permisiion

+++ commands/purge.js
@@ -7,11 +7,12 @@ module.exports = {
         .addIntegerOption(option =>
             option.setName('amount')
                 .setDescription('Amount of message to be cleared.')
-                .setRequired(true))
-        .setDefaultMemberPermissions(PermissionsBitField.Flags.ManageMessages),
+                .setRequired(true)),
 
 	async execute(interaction) {
         const amount = interaction.options.getInteger('amount');
+
+        if(!interaction.member.permissions.has(PermissionsBitField.Flags.ManageMessages)) return interaction.reply({content: 'You do not have permission to use this command'});
 
          await interaction.channel.bulkDelete(amount);
